# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenClaw Mail Secretary API
  version: 2.0.0
  description: >
    Canonical HTTPS REST API surface for clawd-on-desk Electron HUD,
    openclaw-hud-gateway, and the private organization-scoped OpenClaw Cloud Run service.
    The HUD uses only the IAP-protected gateway URL; it never calls the private
    openclaw-api URL directly. Initial OAuth is browser navigation through the
    gateway. The private API performs confidential Google OAuth code exchange
    with the OAuth client secret from Secret Manager, stores the Gmail refresh
    token in Secret Manager, and returns only short-lived Google ID tokens plus
    sanitized claims to the HUD. The HUD never receives, sends, or persists
    refresh tokens, access tokens, authorization codes, code verifiers, or
    OAuth client secrets.
servers:
  - url: https://{gatewayHost}
    variables:
      gatewayHost:
        default: openclaw-hud-gateway.example.run.app
security:
  - googleIdToken: []
...
paths:
...
  /v1/events:
    get:
      tags: [events]
      summary: Poll event inbox for asynchronous HUD updates.
      description: >
        Returns events that have an event_inbox_recipients row for the
        authenticated user's attendee_ref and whose recipient acked_at is null.
        Results are ordered by createdAt ascending (oldest unacknowledged first).
        The cursor is only a paging cursor and does not mark events as consumed.
        HUD restart or cursor loss must not drop unacknowledged events;
        an event becomes fully acknowledged only after every recipient row has
        a non-null acked_at.
      operationId: listEvents
      parameters:
        - name: cursor
          in: query
          required: false
          description: >
            Opaque paging cursor. Implementations should encode the last seen
            createdAt and eventId so ordering is stable even when multiple
            events share the same timestamp. The cursor does not acknowledge
            or consume events.
          schema:
            type: string
            example: "2026-06-03T10:00:00.000Z_evt-00000000-0000-0000-0000-000000000001"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Event inbox page.
          content:
            application/json:
              schema:
                type: object
                required: [events, nextCursor]
                additionalProperties: false
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  nextCursor:
                    type: [string, "null"]
                    example: "2026-06-03T10:00:00.000Z_evt-00000000-0000-0000-0000-000000000001"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/events/{eventId}/ack:
    post:
      tags: [events]
      summary: Acknowledge an event after the HUD displays or processes it.
      description: >
        Idempotently sets acked_at only on the caller's
        event_inbox_recipients row for this event. Duplicate acknowledgements
        from the same caller return 200 with the same eventId. Acknowledging a
        BU-shared event does not hide it from other recipients whose acked_at is
        still null. Returns 403 when no recipient row exists for the caller's
        attendee_ref. Fully acknowledged events are physically deleted only
        after the ACK retention window.
      operationId: acknowledgeEvent
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event acknowledged. Duplicate acknowledgements are idempotent.
          content:
            application/json:
              schema:
                type: object
                required: [success, eventId, acknowledgedAt, fullyAcknowledged]
                additionalProperties: false
                properties:
                  success:
                    type: boolean
                  eventId:
                    type: string
                  acknowledgedAt:
                    type: string
                    format: date-time
                    description: acked_at stored on the caller's event_inbox_recipients row.
                  fullyAcknowledged:
                    type: boolean
                    description: True only when every event_inbox_recipients row for this event has non-null acked_at.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /v1/auth/id-token/refresh:
    post:
      tags: [auth]
      summary: Refresh the HUD short-lived Google ID token without exposing refresh tokens.
      description: >
        Programmatic gateway API used by TokenManager before ID token expiry.
        The HUD sends the current Google ID token in both Proxy-Authorization
        and Authorization. The gateway passes IAP and forwards Authorization to
        the private API while minting a fresh X-Serverless-Authorization token
        with the gateway service account. The private API validates the current
        ID token, uses the Secret Manager Gmail refresh token to obtain a new
        Google ID token, verifies the new claims, and returns only the new
        short-lived ID token plus sanitized claims.
      operationId: refreshHudIdToken
      responses:
        "200":
          description: New short-lived ID token issued to the HUD.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdTokenRefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/auth/bind-subject:
    post:
      tags: [auth]
      summary: Bind a pre-registered BU user on first verified login.
      description: >
        Verifies the Google ID token without requiring an existing subject,
        hashes its normalized email, and atomically binds the subject to exactly
        one pending_subject_bind user. Replays, email mismatches, and already
        bound users fail closed.
      operationId: bindFirstLoginSubject
      responses:
        "201":
          description: Subject bound and user activated.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectBindResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/auth/gmail-self-registration:
    post:
      tags: [auth]
      summary: Register the caller's personal Workspace Gmail account.
      description: >
        Idempotently confirms the personal Workspace Gmail registration already
        provisioned by the confidential OAuth exchange. The exchange validates
        Google claims and BU address conflicts, generates attendee_ref, creates
        the users row, and stores refresh-token/{attendee_ref} before returning
        a usable HUD ID token. Unknown subjects are never admitted through the
        normal application allowlist solely to call this endpoint (D-37).
      operationId: registerPersonalGmailAccount
      responses:
        "201":
          description: Personal Gmail account registered for the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailSelfRegistrationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /internal/auth/oauth-sessions:
    post:
      tags: [internal]
      summary: Create a short-lived OAuth auth session for the gateway.
      description: >
        Gateway-to-private-API internal endpoint. The private API creates
        state and code_verifier, stores them in a short-lived auth session, and
        binds the exact return URI and HUD-provided handoff challenge before
        returning a Google authorization URL. This endpoint is not directly
        callable by the HUD.
      operationId: createOAuthSession
      security:
        - gatewayOidcToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthSessionCreateRequest"
      responses:
        "200":
          description: OAuth auth session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthSessionCreateResponse"
        "401":
          $ref: "#/components/responses/InternalUnauthorized"
        "403":
          $ref: "#/components/responses/InternalForbidden"
  /internal/auth/oauth-exchange:
    post:
      tags: [internal]
      summary: Confidentially exchange Google OAuth code and store refresh token.
      description: >
        Gateway-to-private-API internal endpoint. The private API validates
        state, loads the stored code_verifier, exchanges the authorization code
        at Google's token endpoint using the OAuth client secret from Secret
        Manager, validates ID token claims, stores the Gmail refresh token in
        Secret Manager, stores the short-lived ID token behind the bound
        handoff challenge, and returns only a random one-time handoffCode.
      operationId: exchangeOAuthCode
      security:
        - gatewayOidcToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthExchangeRequest"
      responses:
        "200":
          description: OAuth exchange completed and refresh token stored server-side.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthExchangeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InternalUnauthorized"
        "403":
          $ref: "#/components/responses/InternalForbidden"
  /internal/auth/handoff/redeem:
    post:
      tags: [internal]
      summary: Atomically redeem a handoff code. Invoked only by the gateway.
      operationId: redeemInternalAuthHandoff
      security:
        - bootstrapOidcToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandoffRedeemRequest"
      responses:
        "200":
          description: Handoff consumed and ID token returned to the gateway.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthHandoffResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InternalUnauthorized"
        "403":
          $ref: "#/components/responses/InternalForbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/mail/pending:
    get:
      tags: [mail]
      summary: Poll lightweight pending-mail metadata for HUD display.
      description: >
        Returns metadata only. PII-bearing body, analysis, replyDraft, Calendar
        context, and approvalToken are fetched from GET /v1/mail/{mailId}/detail.
        Claim expiry is enforced by every claim-dependent command, not by this
        read endpoint. Personal mail is visible only to its owner; BU mail is
        visible only to active members of the same normalized BU membership.
      operationId: listPendingMailTickets
      responses:
        "200":
          description: Lightweight pending-mail tickets.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /v1/mail/{mailId}/transfer-targets:
    get:
      tags: [mail]
      summary: List active same-BU members eligible to receive this mail claim.
      operationId: listMailTransferTargets
      parameters:
        - $ref: "#/components/parameters/MailId"
      responses:
        "200":
          description: Pseudonymous transfer targets within the mail resource boundary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferTargetListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    googleIdToken:
      type: http
      scheme: bearer
      bearerFormat: Google ID token
      description: >
        Google-signed ID token returned to the HUD after gateway/API
        confidential OAuth exchange. Sent by the HUD to gateway-facing APIs in
        Authorization, and also in Proxy-Authorization for the first IAP
        programmatic-access approach. The private OpenClaw application validates
        issuer, expiration, aud (the OAuth client ID), hosted domain, and
        allowed subject. Cloud Run URL audience is used only for gateway,
        Scheduler, and Tasks OIDC tokens.
    gatewayOidcToken:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC token
      description: >
        OIDC identity token minted by the openclaw-hud-gateway service account
        for private openclaw-api internal calls. The gateway service account
        must have Cloud Run roles/run.invoker on openclaw-api. The private API
        validates issuer, expiration, audience equal to the openclaw-api Cloud
        Run service URL, and service account email allowlist.
    bootstrapOidcToken:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC token
      description: >
        OIDC identity token minted by the auth-bootstrap service account. The
        private API accepts it only on the internal handoff redeem endpoint.
    adminOidcToken:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC token
      description: >
        OIDC identity token minted by the dedicated admin service account.
        The private API validates issuer, expiration, audience equal to the
        openclaw-api Cloud Run service URL, and exact Admin SA email allowlist.
        Gateway, Scheduler, Tasks, runtime, and end-user tokens are rejected.
    schedulerOidcToken:
      type: http
...